Extract information schema data filtering into a separate class
