+++ core/trino-main/src/main/java/io/trino/connector/informationschema/InformationSchemaPageSource.java
@@ -51,8 +51,8 @@ import static com.google.common.collect.ImmutableMap.toImmutableMap;
 import static com.google.common.collect.ImmutableSet.toImmutableSet;
 import static com.google.common.collect.Sets.union;
 import static io.trino.SystemSessionProperties.isOmitDateTimeTypePrecision;
-import static io.trino.connector.informationschema.InformationSchemaMetadata.defaultPrefixes;
 import static io.trino.connector.informationschema.InformationSchemaMetadata.isTablesEnumeratingTable;
+import static io.trino.connector.informationschema.SystemTableFilter.defaultPrefixes;
 import static io.trino.metadata.MetadataListing.getViews;
 import static io.trino.metadata.MetadataListing.listSchemas;
 import static io.trino.metadata.MetadataListing.listTableColumns;
